+++ tests/projects_test.py
@@ -8,6 +8,8 @@ from mpt.models.user import User
 from mpt.models.project import Project
 from mpt.models.task import Task
 
+from tests.config import TOKEN
+
 class ProjectsTestCase(unittest.TestCase):
     def setUp(self):
         database_name = "mpt_test"
@@ -89,7 +91,7 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_add_new_project(self):
 
-        res = self.client().post('/projects', json = self.new_project)
+        res = self.client().post('/projects', json = self.new_project, headers={'Authorization': TOKEN})
         
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
@@ -100,7 +102,7 @@ class ProjectsTestCase(unittest.TestCase):
 
         self.new_project["manager"] = 0
 
-        res = self.client().post('/projects', json = self.new_project)
+        res = self.client().post('/projects', json = self.new_project, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
         self.assertEqual(data['success'], False)
@@ -110,14 +112,14 @@ class ProjectsTestCase(unittest.TestCase):
 
         self.new_project["manager"] = self.dev_id
 
-        res = self.client().post('/projects', json = self.new_project)
+        res = self.client().post('/projects', json = self.new_project, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
         self.assertEqual(data['success'], False)
         self.assertEqual(data['message'], "Request Cannot Be Processed")
 
     def test_422_when_missing_project_properties(self):
-        res = self.client().post('/projects', json = {})
+        res = self.client().post('/projects', json = {}, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
@@ -126,7 +128,7 @@ class ProjectsTestCase(unittest.TestCase):
     
     def test_405_if_add_project_not_allowed(self):
 
-        res = self.client().post('/projects/20', json = self.new_project)
+        res = self.client().post('/projects/20', json = self.new_project, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 405)
@@ -135,7 +137,7 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_get_projects(self):
 
-        res = self.client().get('/projects')
+        res = self.client().get('/projects', headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -144,7 +146,7 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_get_project_by_id(self):
 
-        res = self.client().get('/projects/{}'.format(self.project_id))
+        res = self.client().get('/projects/{}'.format(self.project_id), headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -153,7 +155,7 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_404_if_project_not_found(self):
 
-        res = self.client().get('/projects/99999')
+        res = self.client().get('/projects/99999', headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -167,7 +169,7 @@ class ProjectsTestCase(unittest.TestCase):
             "manager": self.manager_id,
             "status":"ongoing",
             "assignees": [self.dev_id]
-        })
+        }, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
@@ -176,7 +178,7 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_405_if_update_project_not_allowed(self):
 
-        res = self.client().patch('/projects', json = self.new_project)
+        res = self.client().patch('/projects', json = self.new_project, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 405)
@@ -185,14 +187,14 @@ class ProjectsTestCase(unittest.TestCase):
 
     def test_delete_project(self):
 
-        res = self.client().delete('/projects/{}'.format(self.project_id))
+        res = self.client().delete('/projects/{}'.format(self.project_id), headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_404_if_delete_project_not_exists(self):
-        res = self.client().delete('/projects/0')
+        res = self.client().delete('/projects/0', headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
@@ -200,7 +202,7 @@ class ProjectsTestCase(unittest.TestCase):
         self.assertEqual(data['message'], "Page Not Found")
 
     def test_405_if_delete_project_not_allowed(self):
-        res = self.client().delete('/projects')
+        res = self.client().delete('/projects', headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 405)
@@ -208,7 +210,7 @@ class ProjectsTestCase(unittest.TestCase):
         self.assertEqual(data['message'], "Method Not Allowed")
 
     def test_get_project_tasks(self):
-        res = self.client().get('/projects/{}/tasks'.format(self.project_id))
+        res = self.client().get('/projects/{}/tasks'.format(self.project_id), headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
