+++ tests/tasks_test.py
@@ -9,6 +9,7 @@ from mpt.models.task import Task
 from mpt.models.project import Project
 from mpt.models.activity import Activity
 
+from tests.config import TOKEN
 
 class TasksTestCase(unittest.TestCase):
     def setUp(self):
@@ -96,7 +97,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_add_task(self):
 
-        res = self.client().post('/tasks', json = self.new_task)
+        res = self.client().post('/tasks', json = self.new_task, headers={'Authorization': TOKEN})
         
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
@@ -106,7 +107,7 @@ class TasksTestCase(unittest.TestCase):
   
     def test_error_422_when_project_not_found(self):
         self.new_task["project"] = 0
-        res = self.client().post('/tasks', json = self.new_task)
+        res = self.client().post('/tasks', json = self.new_task, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
@@ -114,7 +115,7 @@ class TasksTestCase(unittest.TestCase):
         self.assertEqual(data['message'], "Request Cannot Be Processed")
 
     def test_error_422_when_missing_task_properties(self):
-        res = self.client().post('/tasks', json = {})
+        res = self.client().post('/tasks', json = {}, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
@@ -123,7 +124,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_get_tasks(self):
 
-        res = self.client().get('/tasks')
+        res = self.client().get('/tasks', headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -132,7 +133,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_get_task_by_id(self):
 
-        res = self.client().get('/tasks/{}'.format(self.task_id))
+        res = self.client().get('/tasks/{}'.format(self.task_id), headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 200)
@@ -141,7 +142,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_404_if_task_not_found(self):
 
-        res = self.client().get('/tasks/0')
+        res = self.client().get('/tasks/0', headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 404)
@@ -150,7 +151,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_405_if_add_task_not_allowed(self):
 
-        res = self.client().post('/tasks/20', json = self.new_task)
+        res = self.client().post('/tasks/20', json = self.new_task, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 405)
@@ -165,7 +166,7 @@ class TasksTestCase(unittest.TestCase):
             "status": "created",
             "assignee": self.dev_id,
             "project": self.new_task["project"]
-        })
+        }, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
@@ -179,7 +180,7 @@ class TasksTestCase(unittest.TestCase):
             "status": "created",
             "assignee": 0,
             "project": self.new_task["project"]
-        })
+        }, headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 422)
@@ -188,7 +189,7 @@ class TasksTestCase(unittest.TestCase):
 
     def test_405_if_update_task_not_allowed(self):
 
-        res = self.client().patch('/tasks', json = self.new_task)
+        res = self.client().patch('/tasks', json = self.new_task, headers={'Authorization': TOKEN})
         data = json.loads(res.data)
 
         self.assertEqual(res.status_code, 405)
@@ -198,14 +199,14 @@ class TasksTestCase(unittest.TestCase):
     
     def test_delete_task(self):
 
-        res = self.client().delete('/tasks/{}'.format(self.task_id))
+        res = self.client().delete('/tasks/{}'.format(self.task_id), headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 200)
         self.assertEqual(data['success'], True)
 
     def test_404_if_delete_task_not_exists(self):
-        res = self.client().delete('/tasks/0')
+        res = self.client().delete('/tasks/0', headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 404)
@@ -213,7 +214,7 @@ class TasksTestCase(unittest.TestCase):
         self.assertEqual(data['message'], "Page Not Found")
 
     def test_405_if_delete_task_not_allowed(self):
-        res = self.client().delete('/tasks')
+        res = self.client().delete('/tasks', headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
         self.assertEqual(res.status_code, 405)
@@ -221,7 +222,7 @@ class TasksTestCase(unittest.TestCase):
         self.assertEqual(data['message'], "Method Not Allowed")
 
     def test_get_task_activities(self):
-        res = self.client().get('/tasks/{}/activities'.format(self.task_id))
+        res = self.client().get('/tasks/{}/activities'.format(self.task_id), headers={'Authorization': TOKEN})
 
         data = json.loads(res.data)
